+++ src/demo/demo.tsx
@@ -1,6 +1,6 @@
 import { h, render, FunctionComponent } from 'preact'
-import { memo } from 'preact/compat'
-import { useCallback, useEffect, useRef, useState } from 'preact/hooks'
+import { memo, useCallback, useEffect, useState } from 'preact/compat'
+import { BrowserRouter as Router, Route, Link } from 'react-router-dom'
 import {
   UIConfigs,
   getInitSdkOptions,
@@ -8,82 +8,14 @@ import {
   getTokenFactoryUrl,
   getToken,
 } from './demoUtils'
-import { BrowserRouter as Router, Route, Link } from 'react-router-dom'
+import SdkMount from './SdkMount'
 
-import { ServerRegions, SdkHandle, SdkOptions } from '~types/sdk'
-
-/*
-The SDK can be consumed either via npm or via global window.
-Via npm there are also two ways, via commonjs require or via ES import.
-In this case we will use the import via the global `window`.
-
-Alternative import styles:
-"commonjs" import style
-const Onfido = require('../index')
-"es" import style
-import * as Onfido from '../index'
-*/
-
-const Onfido = window.Onfido
+import { ServerRegions, SdkOptions } from '~types/sdk'
 
 let port2: MessagePort = null
 let regionCode: ServerRegions = null
 let url: string = null
 const defaultRegion: ServerRegions = 'EU'
-
-const SdkMount: FunctionComponent<{
-  options: SdkOptions | UIConfigs
-}> = ({ options }) => {
-  const [onfidoSdk, setOnfidoSdk] = useState<SdkHandle>(null)
-  const mountEl = useRef(null)
-
-  /**
-   * This side effect should run once after the component mounted,
-   * and should execute the clean-up function when the component unmounts.
-   */
-  useEffect(() => {
-    if (!(options as SdkOptions).mobileFlow) {
-      console.log(
-        '* JWT Factory URL:',
-        url,
-        'for',
-        regionCode,
-        'in',
-        process.env.NODE_ENV
-      )
-    }
-
-    console.log('Calling `Onfido.init` with the following options:', options)
-
-    if (mountEl.current) {
-      const sdk = Onfido.init({
-        ...options,
-        containerEl: mountEl.current,
-      })
-      setOnfidoSdk(sdk)
-
-      window.onfidoSdkHandle = sdk
-    }
-
-    return () => onfidoSdk && onfidoSdk.tearDown()
-  }, [])
-
-  useEffect(() => {
-    if (!onfidoSdk) {
-      return
-    }
-
-    if (options.tearDown) {
-      onfidoSdk.tearDown()
-    } else {
-      onfidoSdk.setOptions(options)
-    }
-  }, [options, onfidoSdk])
-
-  return <div ref={mountEl} />
-}
-
-const SDK = memo(SdkMount)
 
 const SdkDemo: FunctionComponent<{
   hasPreview?: boolean
@@ -139,7 +71,9 @@ const SdkDemo: FunctionComponent<{
           Verify identity
         </button>
       )}
-      {token && <SDK options={options} />}
+      {token && (
+        <SdkMount options={options} regionCode={regionCode} url={url} />
+      )}
     </div>
   )
 }
